+++ lib/src/gregorian/gregorian_date.dart
@@ -75,7 +75,12 @@ class Gregorian implements Date, Comparable<Gregorian> {
 
   /// Create a Gregorian date by using year, month and day
   /// year and month default to 1
-  Gregorian(this.year, [this.month = 1, this.day = 1]);
+  /// [year], [month] and [day] can not be null
+  Gregorian(this.year, [this.month = 1, this.day = 1]) {
+    ArgumentError.checkNotNull(year, 'year');
+    ArgumentError.checkNotNull(month, 'month');
+    ArgumentError.checkNotNull(day, 'day');
+  }
 
   /// Calculates Gregorian and Julian calendar dates from the Julian Day number
   /// [julianDayNumber] for the period since jdn=-34839655 (i.e. the year -100100 of both
@@ -219,11 +224,58 @@ class Gregorian implements Date, Comparable<Gregorian> {
   /// add [days], [months] and [years] separately
   /// note: it does not make any conversion, it simply adds to each field value
   /// for subtracting simple add negative value
+  /// UNSAFE
   Gregorian add({int years = 0, int months = 0, int days = 0}) {
     if (years == 0 && months == 0 && days == 0) {
       return this;
     } else {
       return Gregorian(year + years, month + months, day + days);
+    }
+  }
+
+  /// add [years] to this date
+  /// throws if [years] is null
+  Gregorian addYears(int years) {
+    ArgumentError.checkNotNull(years, 'years');
+
+    if (years == 0) {
+      return this;
+    } else {
+      return Gregorian(year + years, month, day);
+    }
+  }
+
+  /// add [months] to this date
+  /// this Method is safe
+  /// throws if [months] is null
+  Gregorian addMonths(int months) {
+    ArgumentError.checkNotNull(months, 'months');
+
+    if (months == 0) {
+      return this;
+    } else {
+      // this is fast enough, no need for further optimization
+      final int sum = month + months - 1;
+      final int mod = sum % 12;
+      // can not use "sum ~/ 12" directly
+      final int deltaYear = (sum - mod) ~/ 12;
+
+      // todo what to do on leap crash ?
+      return Gregorian(year + deltaYear, mod + 1, day);
+    }
+  }
+
+  /// add [days] to this date
+  /// this Method is safe
+  /// throws if [days] is null
+  Gregorian addDays(int days) {
+    ArgumentError.checkNotNull(days, 'days');
+
+    if (days == 0) {
+      return this;
+    } else {
+      // todo can be simplified ?
+      return Gregorian.fromJulianDayNumber(julianDayNumber + days);
     }
   }